pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "namegen"
version = "0.1.0"
description = "Research-grade realistic name generator with linguistic modeling."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "namegen contributors" }]
keywords = ["names", "generator", "linguistics", "phonotactics", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]
train = [
    # reserved for future neural training extras (torch, transformers, ...)
]

[project.scripts]
namegen = "namegen.cli.main:app"

[project.urls]
Homepage = "https://github.com/nefinfe/namegen"
Source = "https://github.com/nefinfe/namegen"

[tool.hatch.build.targets.wheel]
packages = ["namegen"]

[tool.hatch.build.targets.wheel.force-include]
"namegen/data" = "namegen/data"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
packages = ["namegen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
filterwarnings = ["error"]
